+++ .github/workflows/fmt-check.yml
@@ -10,4 +10,6 @@ jobs:
     steps:
     - uses: actions/checkout@v1
     - name: Build
-      run: cargo fmt --all -- --check
+      run: |
+        cargo fmt --all -- --check
+        cd ctoybox && cargo fmt --all -- --check
