+++ .github/workflows/rust.yml
@@ -30,7 +30,7 @@ jobs:
           python-version: 3.5
       - name: cargo test
         run: |
-          cargo test --release
+          cargo test
       - name: py-install
         run: |
           pip install -r ctoybox/requirements.txt
@@ -40,7 +40,7 @@ jobs:
         working-directory: ./ctoybox
         run: |
           maturin build --release
-          pip install ../target/wheels/ctoybox*.whl
+          pip install ./target/wheels/ctoybox*.whl
       - name: py-unittest
         working-directory: ./ctoybox
         run: |
@@ -49,7 +49,7 @@ jobs:
         name: upload-wheel
         with:
           name: ${{ matrix.platform }}-wheel
-          path: target/wheels/
+          path: ctoybox/target/wheels/
       - name: '[on-create] pypi-publish'
         if: github.event_name == 'release'
         working-directory: ./ctoybox
