+++ ctoybox/.gitignore
@@ -1,1 +1,2 @@
 ctoybox/ctoybox/*
+*.abi3.so
