+++ ctoybox/ctoybox/human_play.py
@@ -1,5 +1,4 @@
-from .ffi import Toybox, Input
-import pygame._numpysurfarray as numpysf
+from . import Toybox, Input
 
 import numpy as np
 import argparse
@@ -7,7 +6,6 @@ import pygame
 import pygame.key
 from pygame.locals import *
 import pygame.surfarray
-import json
 import sys
 
 if __name__ == "__main__":
@@ -80,7 +78,6 @@ if __name__ == "__main__":
             image = tb.get_rgb_frame()
             screen = pygame.display.get_surface()
             img = pygame.surfarray.make_surface(np.swapaxes(image, 0, 1))
-            numpysf.make_surface()
             img2x = pygame.transform.scale(img, dim)
             screen.blit(img2x, dest=(0, 0))
             pygame.display.update()
