--- py_toybox/Cargo.toml
@@ -1,35 +1,0 @@
-[package]
-name = "py_toybox"
-version = "0.1.0"
-authors = ["Emma 'Frank' Tosch <emma.tosch@uvm.edu>", "John Foley <johnf@middlebury.edu>", "Kaleigh Clary <kclary@cs.umass.edu>"]
-readme = "README.md"
-description = "Package containing Rust Toybox games."
-homepage = "https://github.com/toybox-rs/toybox-rs"
-publish = false
-edition = "2018"
-
-[lib]
-name = "py_toybox"
-crate-type = ["cdylib"]
-
-[features]
-default = ["pyo3/extension-module"]
-
-[dependencies]
-serde = "1"
-serde_json = "1"
-serde_derive = "1"
-toybox = {path = "../toybox", version="*"}
-schemars = "0.6"
-pyo3 = { version = "0.13", features = ["abi3-py36"] }
-
-[dependencies.toybox-core]
-version = "*"
-path = "../core"
-
-[package.metadata.maturin]
-requires-dist = ["pillow", "numpy"]
-classifier = [
-  "Programming Language :: Python :: 3.6",
-  "Operating System :: OS Independent",
-]
